backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
pytest>=7.3.1
httpx>=0.24.1
requests>=2.30.0
aiohttp>=3.8.4
pydantic>=2.0.0
pytest-asyncio>=0.21.0
openai>=1.6.0
